packages/db/pyproject.toml
[project]
name = "zion-db"
version = "0.1.0"
description = "Database"
readme = "README.md"
authors = [
    { name = "gwainor", email = "me@gokhan.tr" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.0",
    "python-nanoid>=2.0.0",
    "sqlalchemy>=2.0.46",
    "zion-config",
    "zion-logger",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
zion-logger = { workspace = true }
zion-config = { workspace = true }
